# backend-core/swagger.yaml
openapi: 3.0.2
info:
  title: Boards API
  description: Boards public API documentation
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.17
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /boards:
    post:
      tags:
        - boards
      summary: Create board
      description: Create board
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardObject'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
      security:
        - bearerAuth: []
    get:
      tags:
        - boards
      summary: List owned and shared boards
      description: List boards that a user created and boards the user is a member of
      responses:
        '200':
          description: Successfully listed boards
          content:
            application/json:
              schema:
                type: object
                properties:
                  owned:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardWithUsers'
                  shared:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardWithUsers'
        '400':
          description: Invalid ID supplied
      security:
        - bearerAuth: []
  /boards/{boardID}:
    get:
      tags:
        - boards
      summary: Get board and associated members
      description: Get board and associated members
      responses:
        '200':
          description: Successfully retrieved board with members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardWithUsers'
      security:
        - bearerAuth: []
  /boards/{boardID}/invites:
    post:
      tags:
        - boards
      summary: Create board invites
      description: Create invitations for a specific board
      operationId: createBoardInvites
      parameters:
        - name: boardID
          in: path
          description: ID of the board
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invites:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateInviteObject'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardInvite'
        '400':
          description: Invalid ID supplied
      security:
        - bearerAuth: []
    get:
      tags:
        - boards
      summary: List invites that belong to a board
      description: Returns a list of board invites
      operationId: listInvitesByBoardID
      parameters:
        - name: boardID
          in: path
          description: ID of board
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Status filter
          required: false
          schema:
            type: string
            enum:
              - ACCEPTED
              - PENDING
              - IGNORED
              - CANCELLED
            example: PENDING
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardInvite'
        '400':
          description: Invalid ID supplied
      security:
        - bearerAuth: []
  /invites:
    get:
      tags:
        - invites
      summary: List board invites that belong to a user
      description: Retrieve a list of board invitations.
      parameters:
        - name: status
          in: query
          description: Filter by invitation status
          required: false
          schema:
            type: string
            enum: [ACCEPTED, PENDING, IGNORED, CANCELLED]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardInvite'
  /invites/{inviteID}:
    patch:
      tags:
        - invites
      summary: Update board invitation
      description: Update board invitation belonging to a receiver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - ACCEPTED
                    - IGNORED
                    - CANCELLED
      responses:
        '204':
          description: Successfully updated board invite
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateBoardObject:
      type: object
      properties:
        name:
          type: string
          description: Name of the board
        description:
          type: string
          description: Description of the board
    Board:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: cc143c5c-3a10-46b1-b734-8e2049b719ff
        name:
          type: string
          description: Name of the board
          example: My first board
        description:
          type: string
          description: Description of the board
          example: Isn't this board neat?
        user_id:
          type: string
          format: uuid
          example: d0865843-8494-4d6a-b9be-5c8f7d0e568f
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BoardWithUsers:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: cc143c5c-3a10-46b1-b734-8e2049b719ff
        name:
          type: string
          description: Name of the board
          example: My first board
        description:
          type: string
          description: Description of the board
          example: Isn't this board neat?
        user_id:
          type: string
          format: uuid
          example: d0865843-8494-4d6a-b9be-5c8f7d0e568f
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserWithMembership'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserWithMembership:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID of the user
        name:
          type: string
          maxLength: 255
          description: Name of the user
        email:
          type: string
          format: email
          maxLength: 255
          uniqueItems: true
          description: Email of the user
        membership:
          schema:
            $ref: '#/components/schemas/Membership'
        created_at:
          type: string
          format: date-time
          description: Timestamp indicating the creation date
        updated_at:
          type: string
          format: date-time
          description: Timestamp indicating the last update date
      example:
        id: a5f4c43f-9ae3-4e6d-9682-97351cbb864e
        name: John Doe
        email: johndoe@example.com
        membership:
          role: MEMBER
          added_at: '2023-06-12T10:00:00Z'
          updated_at: '2023-06-12T10:00:00Z'
        created_at: '2023-06-12T10:00:00Z'
        updated_at: '2023-06-12T10:00:00Z'
    Membership:
      type: object
      properties:
        role:
          type: string
          description: Type of membership
        added_at:
          type: string
          format: date-time
          description: Timestamp when user was added to board
        updated_at:
          type: string
          format: date-time
          description: Timestamp indicating the last update date
      example:
        role: MEMBER
        added_at: '2023-05-29T11:30:00Z'
        updated_at: '2023-05-29T11:30:00Z'
    CreateInviteObject:
      type: object
      required:
        - receiver_id
      properties:
        receiver_id:
          type: string
          description: ID of the user to be invited
    BoardInvite:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: cc143c5c-3a10-46b1-b734-8e2049b719ff
        board_id:
          type: string
          format: uuid
          example: b9e95ae4-9c3f-412f-8b3b-201bd7083fc1
        sender_id:
          type: string
          format: uuid
          example: d0865843-8494-4d6a-b9be-5c8f7d0e568f
        receiver_id:
          type: string
          format: uuid
          example: 3c85c78d-d6a2-48c1-aec0-7fe2a9e2d8db
        status:
          type: string
          description: status of board invite
          enum: [ACCEPTED, PENDING, IGNORED]
          example: PENDING
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...